refactor: optimize the structure of files and folders
