+++ src/commands/basic/exit-command.js
@@ -1,4 +1,4 @@
-import { Command } from './command.js';
+import { Command } from '../command.js';
 
 class ExitCommand extends Command {
   async execute() {
