+++ src/commands/file-operation/copy-command.js
@@ -2,7 +2,7 @@ import { join, basename } from 'node:path';
 import { pipeline } from 'node:stream/promises';
 import { createWriteStream, createReadStream } from 'node:fs';
 import { access, stat, mkdir } from 'fs/promises';
-import { Command } from './command.js';
+import { Command } from '../command.js';
 
 class CopyCommand extends Command {
   constructor(currentDir, source, destination) {
